Remove dependency of viewmodel to prototype
